Allow memory region to be supplied for building DefiningMaps
Updated OSRCompilationData::buildDefiningMap to accept a TR_Region
argument to be supplied.  As the validity of DefiningMaps is tied to the
current state of the trees, it's likely that an optimization will want
to build and then discard the DefiningMaps.  Also added an
OSRCompilationData::clearDefiningMap method to allow the client to
discard the DefiningMaps once they are no longer needed.

To simplify this, the OSRMethodData::getDefiningMap method is no longer
responsible for allocating its own DefiningMap - the processing in
OSRCompilationData::buildDefiningMap is now responsible, and once built,
the DefiningMap is added to OSRMethodData via a setDefiningMap method.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>