Expose the restart point in Peephole instruction traversal
We have to do this, at least today, because the Power record form
peephole optimization is currently the only peephole which has a
backwards peephole window. Meaning that it searches for opportunities
by looking back in the instruction stream.

This is problematic because the cross-codegen Peephole optimization
traverses instructions forwards and assumes that peephole windows are
forward facing, such that in the event that a peephole removes the
current instruction being processed we have the previous instruction
processed as a valid point to restart the peephole traversal since the
previous instruction will not have been modified.

The record form peephole violates this assumption in that it has the
power to remove the current and the previous instruction, which means
we no longer have a valid restart point. To fix this issue we expose
the previous instruction as a field and update it in the record form
peephole if the two instructions we are about to remove are consecutive.

This is hopefully a temporary solution until we can support proper
peephole windows and restart points.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>