Added Asserts for large TOC offsets
For cases where the TOC offset is unusually large, I put in an assert instead
of generating instructions. The reasoning was TOC offsets of this size should
never occur so those code paths would never be tested. Also TOC offsets of
that size are likely an indication of a problem elsewhere.

Also changed addConstantToInteger to just truncate since it is handling 32 bit
data which means it is okay for the bits outside the lower 32 bits to be
garbage.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>