Deprecate getJitMethodEntryAlignmentThreshold
Aligning the entry point of a method on a particular boundary is a
performance optimization. However, the use of a threshold introduces
non-determinancy (and arguably randomness run-to-run) into any
performance benefit. Whether or not a particular method is aligned
depends on the methods that have been compiled before it. It also
doesn't take into account the "importance" of any particular method.
For instance, a method invoked millions of times is better suited to be
aligned than one only invoked a few times, but it could be rejected for
alignment because it exceeds the threshold. Unless we align every
method boundary we cannot achieve consistent performance.

As such the benefit of having a threshold is overweighed by the
complexity and non-determinism it introduces and is hence being
deprecated here.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>