AArch64: Create new instruction classes to handle alias instructions
This commit introduces 3 new instruction classes:
- ARM64Trg1ZeroSrc1Instruction
- ARM64ZeroSrc1ImmInstruction
- ARM64ZeroSrc2Instruction

These classes are used to handle alias instructions using the `xzr` register.
For example, `cmp` instruction is an alias of `subs` instruction with `xzr`
as target register.
The current implementation of helper functions for generating alias instructions
such as `generateCompareImmInstruction` allocates a virtual register
and associates it with `xzr` using register dependency.
Ths problem with this approach is:
- That register dependency interferes other register dependencies.
- Those helper functions cannot be used to generate instructions after
 register assignment phase. (generating prologue/epilogue)

The new implementation of those helper functions use newly introduced
instruction classes instead of allocating virtual register for `xzr`.
Those instruction classes does not hold a register for `xzr`.
`generateBinaryEncoding` of those classes encodes register number of `xzr`
to the appropriate position of instructions.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>