Refactor PPCPairedLabelAbsoluteRelocation::apply
Previously, PPCPairedLabelAbsoluteRelocation would use the
updateImmediateField method on the instructions it was going to patch to
set the correct immediates during relocation. However, the
updateImmediateField method uses the old binary encoding system and
contained a lot of handling for instructions that are not meant to be
used as part of a relocation.

To simplify things, PPCPairedLabelAbsoluteRelocation will now make
explicit assumptions about the types of instructions it is passed and
will perform the patching itself.

Signed-off-by: Ben Thomas <ben@benthomas.ca>