Remove the CRLogical instruction flag
Previously, the Power codegen used to have a CRLogical flag that was set
on instructions that performed logical operations on condition register
fields. This was previously only used to abstract away the differences
between these instructions in the old binary encoder. Outside of this
context, abstracting across the entire set of CR logical operations
doesn't make much sense, as they all perform distinct operations and
would thus need special handling anyway. Based on this, this flag is
being removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>