Remove the IsRotateOrShift instruction flag
Previously, the Power codegen had an IsRotateOrShift flag that was set
on instructions representing rotate-and-mask and shift operations. This
was previously only used to control the behaviour of the binary encoder.
Outside of this context, abstracting across all of these operations does
not make sense, as they all have significantly differing behaviour.
Based on this, this flag is being removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>