Switch to using an instruction property for gr0 excludes
Several instructions on Power have an odd quirk where one of their
sources cannot be placed in gr0, since a constant value of 0 will be
used instead. Previously, this was checked for by comparing the
instruction opcode against a set list of instructions with this
property.

Unfortunately, the list of instructions with this property has grown
over time and will likely continue to do so. In order to deal with this,
a new instruction property flag has been added which determines when
this behaviour should apply.

Closes: #5297
Signed-off-by: Ben Thomas <ben@benthomas.ca>