Add a gr0 exclusion for isel extended mnemonics
As it turns out, the isel instruction will use a constant of 0 instead
of the value of gr0 for its RA field. This requires special handling
from the register allocator to ensure that the register being put in the
RA field does not end up in gr0. This behaviour has now been enabled for
all of the isel extended mnemonics.

Signed-off-by: Ben Thomas <ben@benthomas.ca>