Reduce hot loop alignment to 16 bytes on POWER9
Previously, the Power codegen would always align hot loops to 32 bytes
for performance reasons. However, it turns out that this alignment can
be reduced to 16 bytes on POWER9 and later chips without sacrificing
performance. To accommodate this, a new function has been added to get
the preferred hot loop alignment and this alignment will now be used
where necessary.

Signed-off-by: Ben Thomas <ben@benthomas.ca>