Add heap free ratio heuristics for optthruput heap expansion
The expansion and contraction of flat memory for tenured space, used by the optthuput/opavgpause
heap and gencon old gen, is governed by several criteria. Prior to this change, it evaluated a
simple minimum and maximum heap free percentage then checked the GC STW time overhead. If the
current heap free percentage was greater than the configured max free, the heap would not be
expanded even with very high GC time overhead. This change employs a heuristic to offset the
allowable maximum heap free percentage by a factor of the GC time overhead to allow the GC to
expand even with high free memory percentage when the GC time overhead is very high.

For example, with max free memory of 60% and a max CPU overhead of 13%, the heap would not be
expanded when the heap is 70% free even if the GC time overhead is 80%, well beyond the max.
By increasing the maximum allowed free percentage by a factor of the CPU overhead, the heap
would be expanded and the GC time overhead subsequently reduced in such a case.

Minimum free is checked to consider heap contraction. The default Xminf is 30 and the default
Xmaxf is 60. The new offset is as follows:

CPU Overhead	Min Free	Max Free
13%		Xminf+1.69	No heap expansion
20%		Xminf+4		Xmaxf+24
30%		Xminf+9		Xmaxf+54
40%		Xminf+16	Xmaxf+96
50%		Xminf+25	100
60%		Xminf+36	100
70%		Xminf+49	100
80%		Xminf+64	100
90%		Xminf+81	100
100%		100		100

Signed-off-by: Jason Hall <jasonhal@ca.ibm.com>