Run GitHub Actions only on eclipse/omr repository
This prevents GitHub Actions from running on people's forks and every
so often the infrastructure will fail on the GitHub side and an Action
job will report as failed which results in an unnecessary notification
to the forked repo owner. We would like to avoid this, in addition to
wasted resources since these Actions are not meant to be ran on forks.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>