Disable MemoryReference expansion for prefixed instructions
Previously, the MemoryReference expansion logic was being applied to
prefixed loads and stores. This is an issue since this logic assumes
that all displacements are used as 16-bit immediates, which is not true
for prefixed instructions, where the displacement is a 34-bit immediate.
Since this logic is effectively deprecated and should be replaced with
better handling in the locations in the codegen that generate these
loads/stores, all expansion logic (aside from the simple replacement of
NULL base registers with gr0) is being disabled for prefixed
instructions.

Signed-off-by: Ben Thomas <ben@benthomas.ca>