Change formats for plq and pstq
Previously, the plq and pstq instructions in the Power codegen used the
regular RT and RS fields during encoding. While this does produce
correct results, it fails to validate that the register is even. In
order to be able to validate this during binary encoding time, these
instructions have had their formats switch to instead use RTp and RSp
fields.

Signed-off-by: Ben Thomas <ben@benthomas.ca>