Prevent inliner from using a vairant parm to get a call target
Previously, that inliner would assume it was safe to use the type of a
parameter symbol to derive a call target (when the parameter value is
the receiver of a call). Unfortunately, this is not true for variant
parameters because the value written to the parameter slot may be of
a different type than the parameter. This commit corrects the behaviour
by only using the type for symbols of non variant parameter.

Fixes https://github.com/eclipse/openj9/issues/8823

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>