Move code that disables RegDepCopyRemoval
Code that disables the regDepCopyRemoval optimization when value types
are enabled for OpenJ9 was situated under the else branch of an if that
checks "if (!optionSet)".  In other words, the optimization was only
disabled if there were JIT options to be processed.  Fixed by moving
the code out of the if-else.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>