Check whether type for VP constraint is definitely an array
Code in isArrayCompTypeValueType was checking the array component type
for the type associated with a constraint, assuming that it must
represent an array type, without first checking that the constraint
definitely held an array.  That could lead to segmentation faults for
the garbage component type.

Added the missing check to confirm that the constraint is definitely for
an array.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>