Provide a fix for using 64bit instructions in 32bit assembly mode
Either #if or macro can fix it. I chose to go with macro, because
the lxvl dependnecy on [0-7] bit of RB which is not working with

Rest assured that 64bit instructions can run in 32bit mode, since
we only supported 64bit processors.

Temporarily put LEarraycopy back in to not break the build as well.

Signed-off-by: Julian Wang <zlwang@ca.ibm.com>