Implement PC-relative loads/stores via Trg1Imm and Src1
Previously, PC-relative loads and stores could only be encoded by giving
their offset relative to a label. Now, support has been added to specify
the offset directly. While this may seem less than useful since it's not
normally possible to know exact offsets to surrounding instructions due
to e.g. boundary avoidance for prefixed instructions, there is one
particular use case: when the displacement is meant to be patched, it is
necessary to force a displacement of 0 to be used, which cannot be done
using labels.

Signed-off-by: Ben Thomas <ben@benthomas.ca>