Add prefixed load/store exploitation to MR expansion
The MemoryReference::expandInstruction method will now exploit the
prefixed load and store instructions when encountering a displacement
that cannot fit in 16 bits when POWER10 support is enabled. This same
logic will also be applied if the MemoryReference in question has a
label, so it is now possible to load from a label while emitting a
regular load instruction.

Signed-off-by: Ben Thomas <ben@benthomas.ca>