Add a TR::LabelSymbol representing the start PC
Previously, the Power codegen required manual handling for the start PC
of a method, which was determined by the preProcInstruction pointer
after the prologue was generated. As a first step towards eliminating
this special handling and making things more consistent, the codegen
will now emit a label instruction at this location. This should allow
the generalized handling for labels to be used instead in the future.

Signed-off-by: Ben Thomas <ben@benthomas.ca>