Switch to using fixedLoadLabelAddressIntoReg for StartPC
Previously, trying to load the StartPC of the current method would use
loadAddressConstantInSnippet. However, the way in which it did so was
very non-intuitive as it would pass in addi as the load operation and
then expect ConstantDataSnippet to patch the actual StartPC address
instead of the address of that constant in the ConstantDataSnippet. By
switching to instead loading the address of the new label for the
StartPC using the fixedLoadLabelAddressIntoReg helper, all of this
legacy code can be eliminated.

Signed-off-by: Ben Thomas <ben@benthomas.ca>