Use PC-relative loads for ConstantDataSnippet on P10
Previously, loading a value from a ConstantDataSnippet was done by using
fixedSeqMemAccess. This causes a 5 instruction sequence to be used on
64-bit machines to load from the absolute address of the snippet.
However, POWER10 introduces prefixed load instructions that can be used
in a PC-relative addressing mode. ConstantDataSnippet has now been
changed to take advantage of this to perform its loads without needing
to load the absolute address of the snippet.

Signed-off-by: Ben Thomas <ben@benthomas.ca>