Use ConstantDataSnippet for 64-bit constants on POWER10
Previously, the loadConstant helper for 64-bit constants would use the
pTOC where possible. However, on POWER10 the pTOC will be disabled in
favour of using PC-relative loads with ConstantDataSnippet. To reflect
this, the helper in question has been updated to delegate to the
loadAddressConstantInSnippet helper to load from a ConstantDataSnippet
on POWER10 and later.

Signed-off-by: Ben Thomas <ben@benthomas.ca>