Update accessStaticItem to use loadAddressConstant on POWER10
The MemoryReference::accessStaticItem method is in charge of handling
any loads/stores to static addresses. Previously, on 64-bit systems, the
pTOC would be used to load the address of the static to access. However,
under POWER10 the pTOC will be disabled in favour of using
ConstantDataSnippet with PC-relative offsets for these purposes. In line
with this, this method has been updated to use loadAddressConstant when
possible when running on POWER10 or later.

Signed-off-by: Ben Thomas <ben@benthomas.ca>