Move pTOC label processing into processRelocations
When trying to load the address of a label into a register on 64-bit
machines, it is necessary to allocate a pTOC entry to be filled in
later. Previously, such entries were being filled in immediately after
binary encoding, which was fine as long as the label in question was not
part of a snippet.

However, there are several snippets in OpenJ9 which require that their
addresses be loaded into a register. Previously, this was handled
individually by each snippet by patching the mainline code manually with
the correct address. Not only does loading addresses this way cause
needless code duplication and other such issues, it also makes it very
difficult to take advantage of POWER10 PC-relative paddi instructions
in a uniform manner.

In order to prepare for OpenJ9 to be able to handle this in a more
elegant manner, it's necessary to ensure that loading the address of a
label whose code location is set during snippet emission works
correctly. To accomplish this, filling in pTOC entries with label
addresses now occurs during relocation processing.

Signed-off-by: Ben Thomas <ben@benthomas.ca>