Remove InstructionAbsoluteRelocation
Previously, there was a relocation kind for filling in the absolute
address of an instruction. This was completely unused and is also
completely redundant, as one can simply place a label immediately before
or after the instruction in question and relocate using that instead.
For that reason, this relocation kind has been removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>