Reintroduce Dispatcher Rework with fix
Fixes https://github.com/eclipse/openj9/issues/10171

For details regarding the reverted changes being reintroduced here,
please refer back to the original PR
(https://github.com/eclipse/omr/pull/5336). This PR details the issue
with the original change and the patch to fix it.

Initial changes caused intermittent test failures/hangs. This was caused
by a race (introduced by rework) between task dispatch and thread
startup
routine. This occurred when a task was dispatched before a thread
finished its startup routine - before the thread got a chance to enter
its main loop (in worker entryPoint) and "wait for the task". This
caused the thread to miss wakeUp notification and wait indefinitely.

Before the rework, thread status was updated prior to task being
dispatched. Hence, a task dispatched prior to  thread entering the loop
would already have the thread's status as reserved, the thread would
continue on with task rather than wait upon entering the loop. However,
with this new approach, thread status is updated as threads wakes up.
The initial implementation of this approach assumed that a thread must
be "waiting" for a task, which is not the case as a dispatch can occur
before a thread "waits" during thread startup. This means that if a task
is dispatched prior to thread completing startup/entering its loop and
waiting, the thread would simply see it's status as "waiting" and
suspend until  of a task, however the thread would miss the notification
and wait indefinitely in the case described above.

The fix is to simply reorder waiting loop to check wether the thread
should reserve before sleeping.

Signed-off-by: Salman Rana <salman.rana@ibm.com>