Fix bug in postGRA block splitter
In Post GRA Block splitter, there was a bug where for nodes of BCD type,
it was trying to find a global register and in attempt to do so, was
hitting a FATAL_ASSERT. Uncommoning of such nodes should be done using
temp slots with symbol reference containing the correct precision and
length information. This commit removes the FATAL_ASSERT and adds code
to handle such cases.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>