Add a function to break on DF set
Add a function to generate DF flag check in JITCODE and trap when DF is
incorrectly set. Downstream project can add the check to any places they
want.

The motivation of this check is to catch a case where memory can be
corrupted due to incorrect DF flag affecting `rep` instructions. One
case is found in Openj9 but the memory corruption is hard to reproduce
and not all memory corruption can result in a test failure or crash.

However a wrong DF flag should be easier to detect if the check is
executed frequently enough.

Signed-off-by: Liqun Liu <liqunl@ca.ibm.com>