Add assert for use of 64-bit loadConstant on 32-bit
In the Power codegen, there is a loadConstant function for 64-bit
constants that should only be used in 64-bit builds. This function makes
the assumption that sizeof(intptr_t) == sizeof(int64_t) and so will not
work correctly in 32-bit builds. An assert and comment have been added
to this function to clarify this requirement.

Signed-off-by: Ben Thomas <ben@benthomas.ca>