AArch64: Enhance irol evalutor to cancel out double negation
When the OpenJ9 optimizer inlines Integer.rotateRight, it creates
`irol` node with `ineg` child node to adjust shift amount.
AArch64 has instruction for rotate to right only, so we need to negate
the shift amount again. This commit changes `irolEvaluator` to
use child node of `ineg` node as shift amount if the `ineg` node is not commoned.

Also, this commit enhances trace output to print `rorx`/`rorw` instruction
instead of `extrx`/`extrw` instruction.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>