Remove optimization for using a float constant for a dconst
Previously, there was an optimization that would attempt to use a
single-precision floating-point constant when possible in
dconstEvaluator to save space in the pTOC or the ConstantDataSnippet.
However, this is a potentially dangerous optimization and it's not
obvious whether it may suffer from undefined behaviour. In the interests
of safety, this optimization has been removed.

Signed-off-by: Ben Thomas <ben@benthomas.ca>