Add a regression unit test for #5418
The Power codegen's load-hit-store peephole previously contained a bug
where it would assume that any instruction having a load opcode would be
a Trg1Mem instruction and that the last instruction would never be a
store. A new unit test has been added to verify that this issue is
corrected and to prevent future regressions.

Signed-off-by: Ben Thomas <ben@benthomas.ca>