Fix undefined behaviour in the Power LHS peephole
The Power code generator's load-hit-store peephole invokes undefined
behaviour by always assuming that the current instruction is a MemSrc1
instruction and that the next instruction is a Trg1Mem instruction. This
could cause a segfault if the current instruction is the last
instruction or if the next instruction has a load opcode, but is not a
Trg1Mem instruction. This has now been fixed.

Fixes: #5418
Signed-off-by: Ben Thomas <ben@benthomas.ca>