Add use of cmplwi and cmpwi for lcmp* on 32-bit
Previously, the Power codegen implementation of the 64-bit compare
opcodes using register pairs (for 32-bit machines) did not have any
special handling for when the second child was a constant. Code has been
added to make use of the cmplwi and cmpwi instructions where possible to
avoid needing to load the full constant into a temporary register.

Signed-off-by: Ben Thomas <ben@benthomas.ca>