Deprecate label target NOPs on Z
Label target NOPs was a feature on the Z codegen which inserted padding
NOP bytes before labels to align them to cache boundaries. This was an
experimental feature for performance which did not pan out and is
currently unused (disabled by default). We do not envision padding
labels for cache alignment in the future, especially in the way things
are currently implemented as we do now support an ALIGN
pseudo-instruction for this purpose.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>