Fix compiler tests causing lots of warnings
During a previous change, adding the warning flags was moved into
omr_add_executable and omr_add_library. Unfortunately, the OMR compiler
is not yet ready to be compiled with warnings and this causes the
compiler tests to spam lots of warnings. To correct this, these
functions now accept an extra "NOWARNINGS" option that disables this
behaviour, which the OMR compiler uses for its tests and examples.

Fixes: #5296
Signed-off-by: Benjamin Thomas <ben@benthomas.ca>