Remove and replace legacy MemoryReference constructor
Previously, there were two MemoryReference constructors that accepted a
displacement as one of its parameters: one that took it as an int32_t
(a legacy constructor) and one that took it as an int64_t. To ensure
that a call to either one of these constructors was never ambiguous,
an extra int parameter was added at the end of the int64_t constructor.
With this commit, the legacy constructor has been removed and replaced
with the int64_t one, and the extra int parameter has been removed

Signed-off-by: Jackie Midroni <jackie.midroni@mail.utoronto.ca>