Add a flag to indicate codegen support for byteswaps
In order for language frontends to be able to make use of the new
byteswap opcodes, they need a way to determine whether the codegen for
the architecture they're running on supports these opcodes. To allow
this, a new flag has been added to indicate codegen support for
sbyteswap, ibyteswap, and lbyteswap.

Signed-off-by: Ben Thomas <ben@benthomas.ca>