Add pointer to owning metadata in sentinel assumption
This commit facilitates the ability to serialize runtime assumptions.
Any assumptions that have to be deserialized (i.e. reified) is inserted
into the RAT as well as a circular linked list dangled off of the
metadata for the body - this is the paradigm used in OpenJ9 anyway. As
such, this commit adds a field to the sentinel assumption to point to
the owning metadata so that that information can also be serialized.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>