Replace reg dep RealRegister property checks with meaningful queries on Z
Register dependencies rely on "special" RealRegister values to convey
other properties about the register dependency, such as whether a byte
register must be assigned, or whether any assignable register is
suitable for this dependency.  The long-term goal is to not rely on
such special values in the RealRegister enum to convey these properties.
To get there, this commit replaces direct tests of the real register for
queries on the register dependency itself for the property in question.

This commit also fixes a number of compiler warnings on Z where
`RealRegister::RegNum` and `RealRegister::RegDep` enums were compared
directly to test for these properties.  Use a static_cast instead.

The `MayDefine` pseudo real register is never used.  Remove it.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>