Avoid null dereference in inliner due to BCD
The `_callNode` field in the call site may be `NULL` which was not being
checked in code paths guarded by BCD tracing. There was a bug exposed
due to recent changes to enable BCD tracing when `traceCG` is enabled in
#5165. This change activated the BCD tracing code path which exposed the
`NULL` dereference.

The BCD tracing here is actually not very useful because there are
`heuristicTrace` statements immedaitely above which trace pretty much
the same information. Thus the safest and easiest thing to do is to just
remove these extra BCD traces.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>