Fix the badIlOp node generated by RegDepCopyRemoval
For creating a fresh copy for RegDepCopyRemoval optimization, there was
a case when a regLoad is under PassThrough node in which case we would
not have corresponding regStore. To handle this case we were creating a
regStore by getting the type from PassThrough node which would cause
badIlOp. This commit fixes that issue.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>