Fix byteswapped store optimization on Power
Previously, the Power codegen contained an optimization for performing a
store of a byteswapped value in a single instruction. Unfortunately, the
initial implementation of this optimization did not correctly check that
the byteswap node in question was not commoned, which could result in
erratic behaviour. This has now been corrected.

Issue: #5642
Signed-off-by: Ben Thomas <ben@benthomas.ca>