Fix invalid use of stdbrx on P6 and before
Previously, the codegen optimization for performing an lbyteswap and an
lstore in one stdbrx instruction was not properly checking that the
stdbrx instruction was actually available on the current processor.
Since this instruction was only added in P7, this could result in P6 and
earlier processors encountering an illegal instruction.

Signed-off-by: Ben Thomas <ben@benthomas.ca>