Use a default empty message when NULL passed to set error api
`omrerror_set_last_error_with_message` expects a non-null message
though the API doesn't indicate that in the documentation.

Protect against the inadvertant passing of NULL by defaulting to
an empty string for the message.  This ensures that anyone checking
for a message after this call won't get stale data.

Fixes segfaults when the NLS catalog can't be found as it's a common
pattern to pass the NLS result directly to this call.

Signed-off-by: Dan Heidinga <heidinga@redhat.com>