Move Opcodes.enum into OMROpcodes.enum to allow for extensibility
Downstream projects will need to include OMROpcodes.enum and extend the
list inside their own Opcodes.enum. This is to ensure projects which do
not add the root directory containing OMR to the IPATH are able to
include OMR's own list of opcodes. Having OMROpcodes.enum in such cases
prevents circular indludes and name clashes when trying to extend the
opcodes list.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>