Remove enumValue from Opcodes.enum in favor of opcode
The `enumValue` field of this metadata table serves no purpose really
and it is a copy of the `opcode` field, sans the `TR::` suffix. Here we
modify the `opcode` field to remove the `TR::` suffix and instead just
add this suffix in the one location that needs it, i.e. within the
compiler/il/OMRILOpCodesEnum.hpp file.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>