Simplify type_traits guards
We have to support building in environments that
don't support <type_traits> at all, or only support
some of the features we need. Complicating matters
is the fact that not all features have standard test
macros, nor can we rely on the code being built with
CMake (which would allow us to detect compiler
support for specific features at build-config time).
Instead we'll rely on the __cplusplus macro to detect
C++11 and assume that if the compiler supports C++11
it will have most of the traits we need.

An exception to the above is GCC<5, which claims
C++11 support but doesn't have is_trivially_copyable.
For that case, use the OMR_IS_TRIVIALLY_COPYABLE macro
instead, which falls back to a GCC extension where
necessary.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>