Use TR_BitVector instead of CS2::ABitVector in TR_InterferenceGraph
Some of the largest symbols in the compiler obj files originates
from TR_InterferenceGraph resulting from its use of
CS2::ABitVector, while providing no real benefits in terms
of capabilities or performance. Use of CS2::ABitVector has been
replaced with TR_BitVector in TR_InterferenceGraph, which results
in significant reduction in the size of the Compiler component's
obj files, whether it is part of a standalone OMR build or part of
a downstream project.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>