Avoid some reinterpret_casts in CCData ops
Casting to/from data_t* (for non-char data_t) simplifies pointer
arithmetic but requires reinterpret_casts and has potential to
violate aliasing rules. This patch replaces such casts with manual
calculations which allows for some reinterpret_casts to be converted
to safer static_casts.

We can also avoid reinterpret_cast when constructing keys because
keys are now guaranteed to be stored as a sequence of bytes.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>