Add Breakpoint Guard to list of known types for AOT
Breakpoint Guards are used to ensure that if a breakpoint is set on an
inlined method, it is reported to the runtime if the method is executed.
In order to facilate enabling this support for relocatable compilation
(currently only in Eclipse OpenJ9), add TR_BreakPointGuard to the list
of AOT Guard Sites.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>