Cleanup softmx check and apply softmx to nursery expansion
At the moment, heap expansion logic allows nursery size to expand well beyond the softmx limit. Additionally, when the total heap size was above the softmx size, the softmx size was effectively ignored.

This PR fixes both behaviours mentioned above, and makes softmx limit apply to both nursery and tenure (ie, tenure + nursery < softmx).

Additionally, this PR does the following:
- simplified initialization of `desiredContractionFactor` and  `adjustedContractionFactor` in `MemorySubSpaceSemiSpace::checkSubSpaceMemoryPostCollectResize()`. Same idea for `desiredExpansionFactor` and `adjustedExpansionFactor` just above
- refactoring `if(` to `if (`(with space) and `){` to `) {` (with space) in `MemorySubSpaceSemiSpace` and `MemorySubSpaceUnispace`

This PR is a continuation of #5728. Please see it for initial PR feedback on the initial set of changes that was done

Signed-off-by: Cedric Hansen cedric.hansen@ibm.com