Remove inline from CCData static member funcs
Declaring the functions inlined allows the compiler skip emitting
callable symbols for them in `CCData.o`, but because the functions are
declared in a header that's included in other compilation units the
linker complains that there are no implementations for said functions
available, even though they're declared private and not referenced
outside of `CCData.o`.

Unfortunately said functions need to be declared as class members
because they refer to private typedefs in the CCData class.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>